Model-Infra PackingOrder.... Done!!!

+++ BtrGudang.Domain/PackingOrderFeature/QtyType.cs
@@ -2,7 +2,7 @@
 {
     public class QtyType
     {
-        public QtyType(decimal qty, string satuan)
+        public QtyType(int qty, string satuan)
         {
             Qty = qty;
             Satuan = satuan;
@@ -11,7 +11,7 @@
         public static QtyType Default => new QtyType(
             0, "-");
 
-        public decimal Qty { get; private set; }
+        public int Qty { get; private set; }
         public string Satuan { get; private set; }
     }
 }
